Opt: Add Analysis module
The Analysis module provides Def/Use functions to scan a stream of
instructions, code to generate the gen/kill/in/out sets, and a
function to search for a fixpoint on the in/out sets. The gen/kill
sets are not properly initialized yet, so for now the analysis will
not work properly.