Analysis: Store vertex number in data flow sets
The gen/kill/in/out sets are now defined as a set of pairs (n, var)
where n is the number of the vertex where the variable originates
and var is the name of the variable. This is useful because it allows
us to uniquely identify each definition (or use) of a variable and
is necessary to define the gen/kill sets.