Analysis: Initialize gen/kill set for solving dataflow equations
Initialize the gen set of a basic block to all defs in that block.
The kill set of a basic block B is equal to the set of all vars
that share names with vars in gen[B] MINUS gen[B]. I.e. the gen
set is all new bindings introduced and the kill set is all bindings
that can be overwritten by B.

Additionally, this commit hooks up the solver to the initialized
sets for a CFG. As part of this change, we define a "sets_converged"
function that explicitly compares the in/out sets for every vertex,
so that we can correctly tell when the sets have actually converged.