Cfg: Implement multi-vertex removal
Removing multiple vertices from a graph is difficult, since eventually
we will want to remove two connected vertices. Let's say we have four
vertices: A, B, C, D. We also have three edges: AB, BC, CD. If we try
to remove B, we need to merge the AB and BC edges into an AC edge. If
we delete C, we need to merge BC and CD. In order to do both, we have
to be smart enough to realize that AC = BC or CD = BC after one of the
deletions, or else we wouldn't be able to connect A to D at the end.
To do this, we can store a mapping of old vertices to new vertices when
they've been "replaced" after a removal. So in this case, when we remove
B then we get a mapping B -> C. When we remove C we would get a mapping
C -> D. By using this information, we can easily perform the correct
deletion because we're able to update all the edges to have the new
source/destination rather than using an old vertex that's been deleted.