Analysis: Revert use[v] to return a list
Previously, I had changed use[v] to return a VSet.t. This implied
that the vertex number was important information, as a VSet.t will
always include a vertex number. This is highly confusing, since a
user might be inclined to try performing set operations using this
use[v] set. However, since it essentially "invents" its own defs
from the vertex numbers, a union with this set would add a lot of
defs that don't actually exist. Instead, we should avoid returning
any of this extra information because there is no real place where
the vertex numbers of a use would be useful.