Cfg: Add hashtbl_of_cfg; support removal of vertices without any successors
Cfg.hashtbl_of_cfg allows us to create a hash table mapping of vertex IDs
to the actual vertices. This is useful because we can quickly look up the
vertex that a def belongs to. Otherwise, we would have to iterate over all
of the vertices and find a matching vertex number which is much more
computationally intensive.

Previously there was a bug if a vertex had no successors, where we would go
to the fallthrough case thinking it had more than 1 successor. This was an
issue for the final instruction in a program if it was dead-code, because
it wouldn't match against the [succ] case.