Cfg: Add first_instr function to find first instruction in the CFG
When we delete the first instruction in the CFG, we lose our init
node. To replace it, we can simply go thru the vertices and find
the vertex with the lowest number, which will be our new node. This
works because if all the code above it was dead then there is no
path of execution that would get us beyond that point (since that
would require a branch or jump, which are critical).