Analysis: Make it work
Four big fixes:

1. Fixpoint convergence is now properly tested for

Previously, we were calling fixpoint when it DID converge. Quick
fix to call it when it does NOT converge, so that we keep iterating.

2. Fixpoint now passes in the updated sets

Previously, we were passing in the old sets in fixpoint, so that
fixpoint would never actually converge, because it would always
restart on the same initial input.

3. collect_dead_code now checks for a mark before adding to worklist

Previously, we were not checking whether a vertex had been marked
before adding it to the work list in collect_dead_code. This is an
issue because in a loop we may try to queue the same vertex over and
over. The easy fix is to just check whether it has been marked, since
we already have that info.

4. ident_match now uses value equality instead of physical equality

Previously, we were using the == operator in ident_match, which would
compare the pointers of the strings rather than their values. This is
fine if all the strings we compare are interned, but since they're
usually coming from the parser they will have different pointers
(they were created dynamically & the OCaml string module doesn't do
any auto-interning). So instead, we just do a deep comparison and make
sure that the strings' values match. This now allows us to match against
ALL of the matching strings, rather than only the ones from the exact
same def.

And now it works.