Parser: Store lexed numbers as string until parsing phase
This is a quick fix to make floating point numbers parse correctly.
Previously a number in the form 0.0001 would be parsed as two ints
and a DOT in between, resulting in [INT 0, DOT, INT 1]. This is
obviously incorrect because it removes the left-padded 0s for the
decimal, which results in an entirely different number being parsed.
Instead of parsing the integer immediately, we can simply store the
string representation and allow the parser to convert the string to
the proper type of number.