test.sh: Optimize all example files
Added test.sh, which runs the optimize code on all of the *.ir
files in the examples directory and stores their output in the
out directory in a file with the same name. This is useful for
testing, since it allows us to quickly run all of the files
after we make changes. We can use the generated optimized ir to
verify that the optimizer is running as expected after we make
changes.