run.sh: Renamed test.sh to run.sh
Renamed test.sh to run.sh to make naming clearer. I decided to
do this while creating a test script since it will be clearer
to name the test file test.sh and it is clearer to change the
name of the run file to run.sh since it is running all the files,
and doesn't actually test anything.