test.sh and expected/: Added a test script
Stored the expected optimized values of the example *.ir files
in the expected directory, which is designed to store expected
outputs for testing.

Added the test.sh file, runs all the examples and compares their
results with the expected outputs.

test.sh is designed to be used to quickly test the optimizer after
making changes to make sure they aren't causing problems.