Run/Test: Terminate when command fails
Updated the run.sh and test.sh scripts to exit whenever a command
in them fails by setting the "set -e" command. Previously, even
if the code didn't compile, when you ran test it would still output
results, but since the code didn't run the results were the same as
the last successful run, which could be confusing. Also, if there was
a compilation error, it would be printed repeatedly each time the diff
command was run. Now, if the code doesn't compile the error is printed
once and the script exits.