Analysis/Dead/Optimize: Modularize the data flow equation solver
Analysis: Added the module type Dataflow, which declares functions
that are used to implement the data flow solver. Moved the data flow
code to the Make module, and modified them to work with any module
that implements Dataflow.

Dead: Implemented all the functions in Dataflow and helper functions.
Many of these function implementations were copied directly from the
old Analysis. The helper functions were directly copied and the
implemented functions were taken out of the older Analysis init and
solve functions.

Optimize: Updated the code that calls data flow functions to work
with the changes to Analysis and Dead.