Copy/Optimize: Create Dataflow module for copy propagation and implement Dataflow functions.
Created the Copy module for copy propagation. This function
implements all the items in the Dataflow signature. The kill
set must include not just defs that change the destination, but
also defs that change the assignment's rval.

In order to do this, I added more information to type,
specifically for `assign`. If the added bool is false,
then the last 2 parts of the type can be ignored. When
it is true, it is an `assign` and passes the line number, dst,
and rval. The solver's traversal then uses this additional
info to check if anything in the kill set changes either the
dst or rval.

The dataflow equation for the in set is the intersection of
the predecessor's out sets. Other than these changes, the
Copy module is identical to the Dead module.

I commented out the dead code elimination code in optimize and
replaced it with code testing the dataflow equations for Copy.