Copy: When the current BB is a def, add all assignments x := y, where x or y is the rval of the current def
Moved the search for rvals in assigns being changed from current def
from the solve_traverse function to the init_fold function. This was
accomplished by replacing the defs function with a function that
filters all the defs to only assigns with x := y, where x or y is
being redefined.