Create a fuzzer for testing the parser/optimizations
This implements a fuzzer that generates custom programs so that we can
test our optimizer on randomized tests. The goal here is to find edge
cases that we can't think of, but the compiler is able to create. Note
that for now the fuzzer only works on the main function (so no args or
return values) and doesn't support function calls -- this stuff is just
kinda hard to add and test, so I'll leave it out unless it seems super
important. However, that stuff will never get deleted by deadcode since
function decls, returns, and calls are all critical ops.