Format: Add new line at end every function
Previously, newlines weren't added at the end of the file because the
main function didn't have a concat string after it. The parser currently
doesn't support programs that end without a main function, so this is
a quick fix to get fuzzed/optimized programs re-parsed. However, we should
try to clean up the parser so that it doesn't have this restriction.