CFG: Add update_vertices
This function is used for copy propagation, where we want to modify
vertices. The challenge here is that vertices are immutable, so we
can't just tell OCamlgraph to update a vertex's label. Instead, we
need to schedule a list of changes to make to the graph and apply
those in the update_vertices method. This works similar to how we
remove vertices in deadcode elimination: we keep track of each
vertex's replacement and when we update the edges in the graph, we
copy that original edge to the replaced vertices. For example, if
we have a vertex A and a vertex B and we want to replace A with A'
and B with B': simply adding A' => B and A => B' edges won't work,
since these will disappear as soon as we remove the old vertices.
Instead, we need to use the mappings to apply this change to the
final graph, so that we add the edge A' => B'.