Copy/Cfg/Optimize: Replace variables in operands using copy propagation
Copy:
	Added the copy_prop function, which is an in-place implementation
of copy propagation. It works by iterating over every vertex and getting
every operand in every instruction. It then recursively maps the operands
to their copy prop value, and if anything was changed during this process
the vertex containing the instruction is updated.

Cfg:
	2 unrelated changes were made here. First of all, we added a new
function, update_vertices, which replaces a vertex with another vertex.
It creates new edges from the original vertex's preds to the new vertex,
and from the new vertex to the original vertex's succs, then removes
the old vertex.

	Also, on a seperate note I added a new edge type, `Unreachable.
An unreachable edge goes from a Goto or Return to the next instruction,
if one exists. This is to fix a bug in get_code, where the instruction
after a Goto or Return would be skipped if it was not a label that was
pointed to. get_code treats unreachable edges the same way it treats
fallthrough edges.

Optimize:
	Was changed to test copy propagation.