Copy/Analysis/Optimize: Fixed a bug in the copy propagation dataflow sets where during the intersection of 2 out sets, if they all include the same mapping for an assignment with the same variables but the assignments occured on different lines, they would not be put into the next in set.
For example in the program
    x := y
    br label1 ___
    x := y
label 1: print x

no matter how the branch computes x will be assigned to y at the print
statement. Previously, the print statement's in set would think that the
x := y assignments on lines 1 and 3 are different since they have different
line ids. This is incorrect behaviour since in both cases print x -> print y
for all possible values of y. Now it treats them as the same so this mapping
to print y will occur.

This was done by changing the compare to ignore the vertex ids in each element.