Driver: Handle exceptions in iterate
If an exception occurs during iteration, we will calculate the first
instruction of the current CFG & return it. Since the CFG is mutable
there's no need to return the current CFG or anything like that. Note
that there are cases where the CFG could become "corrupt" if we didn't
apply optimizations correctly or completely, but these are very unlikely.